+++ src/main/java/com/atwoz/global/config/async/AsyncConfig.java
@@ -1,7 +1,9 @@
-package com.atwoz.global.config;
+package com.atwoz.global.config.async;
 
+import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.AsyncConfigurer;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
@@ -10,7 +12,7 @@ import java.util.concurrent.ThreadPoolExecutor;
 
 @EnableAsync
 @Configuration
-public class AsyncConfig {
+public class AsyncConfig implements AsyncConfigurer {
 
     private static final String ASYNC_EXECUTOR = "asyncExecutor";
     private static final String ALERT_CALLBACK_EXECUTOR = "alertCallbackExecutor";
@@ -44,5 +46,10 @@ public class AsyncConfig {
         executor.setWaitForTasksToCompleteOnShutdown(WAIT_TASK_COMPLETE);
         executor.initialize();
         return executor;
+    }
+
+    @Override
+    public AsyncUncaughtExceptionHandler getAsyncUncaughtExceptionHandler() {
+        return new AsyncExceptionHandler();
     }
 }
